chore(user-module) : rename file name
