+++ src/modules/students/students.module.ts
@@ -4,7 +4,7 @@ import { Student } from 'src/entities/student.entity';
 import { StudentsRepository } from './students.repository';
 import { StudentsService } from './students.service';
 import { UsersModule } from '../users/users.module';
-import { StudentsController } from './student.controller';
+import { StudentsController } from './students.controller';
 
 @Module({
   imports: [UsersModule, TypeOrmModule.forFeature([Student])],
