+++ src/modules/teachers/teachers.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Teacher } from 'src/entities/teacher.entity';
 import { UsersModule } from '../users/users.module';
-import { TeachersService } from './teacher.service';
 import { TeachersController } from './teachers.controller';
 import { TeachersRepository } from './teachers.repository';
+import { TeachersService } from './teachers.service';
 
 @Module({
   imports: [UsersModule, TypeOrmModule.forFeature([Teacher])],
